move str_to_bool_into Utils
